Se terminaron las pruebas de Package, tambien se corrigieron temas del models, services y middlewares para que pasen los test

+++ models/package.js
@@ -22,11 +22,11 @@ Package.init(
       allowNull: false,
     },
     quantityOfPackages: {
-      type: S.DECIMAL,
+      type: S.INTEGER,
       allowNull: false,
     },
     quantityInOrders: {
-      type: S.DECIMAL,
+      type: S.INTEGER,
       defaultValue: 0,
     },
     deleted: {
